# Optimize grapheme split method