I have an app with a dictionary of several thousand words I want to preprocess. At current, this library is the fastest and most correct grapheme splitter I've found on npm, but I still was interested in seeing if we could get it to a speed where a user wouldn't notice any lag when my app preprocesses all the words. I managed to cut runtime (tested on a list of 8000 words, containing various amounts of conjoining unicode) down to roughly a quarter using a few optimizations in the outermost split method.

The first optimization I employed was to avoid the unnecessary allocation of a new array that isn't ultimately used. `[...str].map` allocated and filled a new array with the individual unicode characters, then immediately allocates and fills another array with the characters mapped to code points and discards the first array. I switched this to a `for(const char of str) codePoints.push(char.codePointAt(0) ...` to avoid the cost of the extra array allocation. 

The next optimization I found was to avoid the round trip between `codePointAt` and `fromCodePoint`. By keeping track of a map (from codePoint indices to original string indices), I could simply call `str.split` and pass the mapped indices to get a substring, and skip the translation and concatenation in `fromCodePoint`.

A last optimization I was able to find was switching from `for..of` to the `while (i < str.length)` you see in the final code. To avoid requesting the computation of each codePoint twice (once during iteration, and then once more in `codePointAt(0)`), I changed the loop to call `str.codePointAt(i)`. I could keep track of the size of each unicode character by determining whether the codePoint was > 65535 (0b1111111111111111).

Again, all told this brought preprocessing times down to ~25% what they were before. I hope these optimizations meet you well.