@orisano 2019-02-19
sorry for late. thanks your PR.
can you add benchmark code?