@orisano 2019-02-20
on my local environment.
Before:
> node benchmark.js
simple string: 19.76638799905777ms
complex string: 49.703597992658615ms

After:
> node benchmark.js
simple string: 5.424661993980408ms
complex string: 16.687121003866196ms

thanks for your optimization!!!